agents/ten_packages/extension/azure_vision_python/requirements.txt
azure-ai-vision-imageanalysis